Доделал ежедневник

+++ my-diary/my_diary/__init__.py
@@ -10,10 +10,11 @@ def action_show_tasks():
     with get_connection() as conn:
         tasks = storage.show_tasks(conn)
 
-    template = 'Задача: {} Описание: {} Подробное описание: {} Статус: {} Время выполнения: {}'
+    template = 'Задача: {} Описание: {} Подробное описание: {} Статус: {} Время выполнения: {}\n'
 
     for task in tasks:
         print(template.format(task[0], task[1], task[2], task[3], task[4]))
+
 
 def action_add_task():
     task = input('\nВведите задачу: ')
@@ -26,29 +27,44 @@ def action_add_task():
     with get_connection() as conn:
         storage.add_task(conn, task, task_description, data)
 
-
 def action_edit_task():
-
-    id = input('Введите номер задачи: ')
-    edited_task = input('Введите исправление задачи: ')
     with get_connection() as conn:
-        storage.edit_task(conn, edited_task, id)
-
+        id = int(input('Введите номер задачи: '))
+        if id in storage.test_id(conn):
+            edited_task = input('Введите исправление задачи: ')
+            edited_task_description = input('Введите исправление описания: ')
+            storage.edit_task(conn, edited_task, edited_task_description, id)
+        else:
+            print('Не верно введен номер задачи')
 
 def action_end_task():
-    done_id = int(input('Введите номер завершенной задачи: '))
     with get_connection() as conn:
-        storage.end_task(conn, done_id)
+        done_id = int(input('Введите номер завершенной задачи: '))
+        if done_id in storage.test_id(conn):
+            storage.end_task(conn, done_id)
+        else:
+            print('Не верно введен номер задачи')
 
 def action_repeat_task():
-    repeat_id = int(input('Введите номер задачи которую нужно повторить: '))
     with get_connection() as conn:
-        storage.repeat_task(conn, repeat_id)
+        repeat_id = int(input('Введите номер задачи которую нужно повторить: '))
+        if repeat_id in storage.test_id(conn):
+            storage.repeat_task(conn, repeat_id)
+        else:
+            print('Не верно введен номер задачи')
 
 def action_delete_task():
-    id = input('Введите номер задачи для удаления: ')
     with get_connection() as conn:
-        storage.delete_task(conn, id)
+        id = int(input('Введите номер задачи для удаления: '))
+        if id in storage.test_id(conn):
+            storage.delete_task(conn, id)
+        else:
+            print('Не верно введен номер задачи')
+
+def action_test_id():
+    proverka = int(input())
+    with get_connection() as conn:
+        print(storage.test_id(conn))
 
 def action_show_menu():
 
@@ -59,6 +75,7 @@ def action_show_menu():
 4. Завершить задачу
 5. Начать задачу сначала
 6. Удалить задачу
+7. Тест айди
 m. Показать меню
 q. Выйти
 ''')
@@ -81,6 +98,7 @@ def main():
         '4': action_end_task,
         '5': action_repeat_task,
         '6': action_delete_task,
+        '7': action_test_id,
         'm': action_show_menu,
         'q': action_exit,
     }
